backend/requirements.txt
fastapi 
uvicorn 
faiss-cpu
sentence-transformers 
pydantic
mangum